Data cleaning and Visualization
My first experience with R, the purpose of this project was to experiment with a text file containing group chat history retrieved from the GroupMe API in HTML format. Initially the information had to be cleaned into a workable format, and was subsequently used to analyze messenger frequency, average message length per user, and to measure activity in the group chat based on the time of day.